requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
ldap3==2.9.1
sqlalchemy==2.0.25
pandas==2.2.0
openpyxl==3.1.2
schedule==1.2.1
slowapi==0.1.9
requests==2.31.0
